refactor(proof): dto 디렉토리 변경으로 인한 의존성 path 변경
